remove runInSeparateProcess and use different files
